fix: moved all logic to pages/Todo and updated exports
