--- react-basics/solutions/todoApp/src/components/TodoForm/TodoForm.types.ts
@@ -1,6 +1,0 @@
-import { CreateTodoSchema, Todo } from "../../types/todo";
-
-export type TodoFormProps = {
-  todos: Todo[];
-  addTodo: (data: CreateTodoSchema) => void;
-};
