--- react-basics/solutions/todoApp/src/components/TodoForm/index.ts
@@ -1,3 +1,0 @@
-import { TodoForm } from "./TodoForm";
-
-export { TodoForm };
