--- react-basics/solutions/todoApp/src/components/TodoItem/index.ts
@@ -1,3 +1,0 @@
-import { TodoItem } from "./TodoItem";
-
-export { TodoItem };
