--- react-basics/solutions/todoApp/src/components/TodoList/index.ts
@@ -1,3 +1,0 @@
-import { TodoList } from "./TodoList";
-
-export { TodoList };
