--- react-basics/solutions/todoApp/src/components/TodoMessage/index.ts
@@ -1,3 +1,0 @@
-import { TodoMessage } from "./TodoMessage";
-
-export { TodoMessage };
