+++ react-basics/solutions/todoApp/src/pages/Todo/Todo.types.ts
@@ -1,7 +1,7 @@
 import { z } from "zod";
-import { createTodoSchema } from "../schemas/createTodoSchema";
+import { createTodoSchema } from "./schemas";
 
-export type Todo = {
+export type TodoModel = {
   id: string;
   name: string;
   completed: boolean;
