+++ react-basics/solutions/todoApp/src/pages/Todo/components/TodoForm/TodoForm.types.ts
@@ -1,0 +1,6 @@
+import { CreateTodoSchema, TodoModel } from "../../Todo.types";
+
+export type TodoFormProps = {
+  todos: TodoModel[];
+  addTodo: (data: CreateTodoSchema) => void;
+};
