+++ react-basics/solutions/todoApp/src/pages/Todo/components/TodoItem/index.ts
@@ -1,0 +1,1 @@
+export { TodoItem } from "./TodoItem";
