+++ react-basics/solutions/todoApp/src/pages/Todo/components/TodoList/index.ts
@@ -1,0 +1,1 @@
+export { TodoList } from "./TodoList";
