+++ react-basics/solutions/todoApp/src/pages/Todo/components/TodoMessage/index.ts
@@ -1,0 +1,1 @@
+export { TodoMessage } from "./TodoMessage";
