+++ react-basics/solutions/todoApp/src/pages/Todo/index.ts
@@ -1,0 +1,2 @@
+export { Todo } from "./Todo";
+export { type TodoModel, type CreateTodoSchema } from "./Todo.types";
